fixing FetchApi
