+++ src/components/atoms/track-0/InfoTracker.jsx
@@ -1,39 +1,26 @@
-
-import { useState, useEffect } from 'react';
-import fetchData from '../../../services/fetchData';
+import FetchData from '../../../services/FetchData';
 import './infoTracker.css';
 
 const InfoTracker = () => {
-    const [apiData, setApiData] = useState(null);
-    const [error, setError] = useState(null);
+  const apiEndpoint = 'all';
+  const { data, error } = FetchData(apiEndpoint);
 
-    const apiEndpoint = 'all'; 
-
-    useEffect(() => {
-        fetchData(apiEndpoint) 
-        .then((api) => {
-            setApiData(api);
-        })
-        .catch((error) => {
-            setError(error);
-        });
-    }, [apiEndpoint]); 
-
-    return (
-        <>
-            {error ? (
-                <h1>Error: {error.message}</h1>
-            ) : (
-                apiData && 
-                <div className='infoTracker-card'>
-                    <div><h4>{apiData.cases}</h4></div>
-                    <div><h4>{apiData.deaths} </h4></div>
-                    <div><h4>{apiData.todayDeaths} </h4></div>
-
-                </div>
-            )}
-        </>
-    );
+  return (
+    <>
+      {error ? (
+        <h1>Error: {error.message}</h1>
+      ) : (
+        data && (
+          <div className='infoTracker-card'>
+            <div><h4>{data.cases}</h4></div>
+            <div><h4>{data.deaths} </h4></div>
+            <div><h4>{data.todayDeaths} </h4></div>
+            <div><h4>{data.todayDeaths} </h4></div>
+          </div>
+        )
+      )}
+    </>
+  );
 };
 
 export default InfoTracker;
