+++ src/components/molecules/Navbar.jsx
@@ -1,5 +1,5 @@
 import "./navbar.css"
-import coronavirus from "../../assets/img/coronavirus.png"
+import coronavirus from "../../../assets/img/coronavirus.png"
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'bootstrap-icons/font/bootstrap-icons.css';
