+++ src/components/molecules/tracker0/Tracker0.jsx
@@ -1,5 +1,5 @@
 import InfoTracker from "../../atoms/track-0/InfoTracker";
-import Top10 from "../../atoms/Top10";
+import Top10 from "../../atoms/track-0/Top10";
 
 const Tracker0 = () => {
   return (
